requirements.txt
asgiref==3.8.1
boto3==1.34.131
botocore==1.34.131
Django==5.0.6
django-dotenv==1.4.2
django-storages==1.14.3
djangorestframework==3.15.2
jmespath==1.0.1
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
